Add extra code samples; add extra "Add" method for config key components to make ML.Components using unnecessary.
